+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 - Added support for Spanish language (Thanks [ssergio-ll](https://www.github.com/ssergio-ll))
+- Ability to set the battery charge limit from 30% to 100% in steps of 1
+- Added application icon to system tray
 - Some optimizations
 
 ## [0.2.0] 2022-10-16